Added a php module for php-based interfaces Simply include utils/jucs.php in your program and use the submit function.
minor changes:
	daemon, core, update: disabled MySQL update when DB_HOST is not defined.
	daemon, jfsck: daemon now uses jfsck to terminate.
	update: fixed typo in table name.

+++ utils/jfsck.sh
@@ -1,10 +1,17 @@
 #!/bin/sh
 
+set -e
+
 exec 2>/dev/null
 
-kill `cat $JUDGE_PREFIX/judge.pid`
+[ -e $JUDGE_PREFIX/judge.pid ] && kill `cat $JUDGE_PREFIX/judge.pid`
 sleep 1
-kill -s 9 `cat /var/run/judgeinit.lock0`
-kill -s 9 `cat $JUDGE_PREFIX/judge.pid`
+[ -e $JUDGE_PREFIX/init.pid  ] && kill -s 9 `cat $JUDGE_PREFIX/init.pid`
+[ -e $JUDGE_PREFIX/judge.pid ] && kill -s 9 `cat $JUDGE_PREFIX/judge.pid`
+
+rm -f $JUDGE_PREFIX/init.pid
+rm -f $JUDGE_PREFIX/judge.pid
+
+rm -f $FIFO $LOCK
 
 echo hasta la vista, baby! :-t
